requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
rank-bm25
anthropic
python-dotenv
requests
